Changed register access API to always pass values as uint64_t.

+++ src/proc.h
@@ -106,13 +106,13 @@ reg_info_t *sim_get_register_info (sim_t *sim,
 bool sim_read_register (sim_t   *sim,
 			int     reg_num,
 			int     index,
-			uint8_t *val);
+			uint64_t *val);
 
 // returns false if reg_num or index out of range
 bool sim_write_register (sim_t   *sim,
 			 int     reg_num,
 			 int     index,
-			 uint8_t *val);
+			 uint64_t *val);
 
 void sim_press_key (sim_t *sim,
 		    int keycode);
